backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.3
python-multipart==0.0.9
langchain==0.2.12
langchain-openai==0.1.8
pypdf==4.2.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic-settings==2.3.4
